Move to generic hsmp interface
The current code exports a group of hsmp1_[get|set]_* routines
that are used to get and set tunables. Instead of naming these
hsmp1_XXX they should just be hsmp_XXX (drop the '1'). Callers
of these routines do not, and should not, need to worry about
the mechanism to get/set the data. They just need the work done.

This patch changes the naming to drop the '1' from the name. If
and/or when we need to move additional mechanisms to get/set the
data, this generic interface can determine which mechanism to use.

This patch willalso move some of the generic pr_* calls from the
sysfs code to the hsmp_XXX routines.

Sifgned-off-by: Nathan Fontenot <nafonten@amd.com>