Define SMU Ports
Introduce the notion of an SMU port for accessing PCI config space.
This allows the driver to use the smu_pci_[read|write] routines to
access PCI config space for HSMP access and for general purpose
access.

This patch is based on a patch from Lewis Carroll that used ports
for smu PCI access.

Signed-off-by: Nathan Fontenot <nafonten@amd.com>
Signed-off-by: Lewis Carroll <lewis.carroll@amd.com>