From: Lewis Carroll <lewis.carroll@amd.com>
Update xgmi2_width interface

Update the sysfs interface to only take a single parameter for the xgmi
width. This will result in the min/max values being set to the same value
when making a change through sysfs. The same valid values are still
enforced.

This patch also adds the ability to pass a value of -1 for the xgmi width.
A value of -1 will allow for autonomous link width settings. With
autonomous link width settings enabled on values of x8 and x16 will be
selected.

Signed-off-by: Lewis Carroll <lewis.carroll@amd.com>
Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>