amd_hsmp.c: Fix logic error in NBIO enumeration
In the function f17h_m30h_init, a recent change broke the intended
qualifier for the while loop. To make the code clearer, separate out
the test for null dev (which signifies the end of walking the device
list) from the count of found NBIOs (the test is intended to prevent
array out of bounds, which could only occur in the case of an
unexpected error walking the PCI device list).

Also updated some comments and fixed some spacing

Signed-off-by: Lewis Carroll <lewis.carroll@amd.com>