Add verbose option to test script
To reduce the amount of output from the test script this updates
the script to limit normal output to the broad areas being tested
and their pass/fail status.

Using the verbose option, -v, to the script will provide verbose
output including the pass/fail status for every test.

This update also renames pr_debug() to pr_verbose() to reflect that
the output from it is comntrolled by the verbose option.

Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>