Correct hsmp driver loading for testing
The current test script behavior assumes the hsmp driver
is not loaded and makes an attempt to load it. For systems
that pre-load the driver this is incorrect.

To correct this the logic used to look for and load a hsmp driver
for testing will follow the policy.

1) If a hsmp driver is already loaded, the test script will use the
already loaded driver to do testing.

2) If a hsmp driver module is specified with the (new) -k option,
the test script will attempt to load this driver and use it for
testing.

3) If a hsmp driver module is not specified the test script will
look for a hsmp driver module (amd_hsmp.ko) in the current
directory to load and use. This maintains current functionality.

The test script will now also only unload the hsmp driver if it
loaded one, a pre-loaded driver will not be removed.

Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>