Make hsmp test module loading follow hsmp module logic.
Update the logic for loading and using the hsmp test kernel
module to follow what is done for testing the hsmp driver.

This patch adds a '-t' option to specify the hsmp test module.

Additionally, an update is made so that testing can occur even
when a hsmp test module isn'tloaded or fails to load. A flag
is set to allow skipping of hsmp test module testing.

Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>