Add union for properly accessing SMU firmware version
Current source does not tell the compiler everything that is needed for
storing the SMU firmware version as a u32 (as read back from the HSMP
message to get the firmware version) and accessing the SMU firmware
version as individual bytes. Add a union to store this as u32 and access
this as a struct with four u8s. No functional change just proper coding.

Signed-off-by: Lewis Carroll <lewis.carroll@amd.com>