Fix bus to NBIO discovery
Per previous concerns, use of DstFabricID in the CfgAddressMap registers
is subject to incorrect mapping due to an MCTP work-around on Rome.
Modify the code to use a new discovery method as documented in the code
comments in the patch.

This patch also changes the way PCI busses are presented to the user in
sysfs - instead of only presenting the bus on which the IOHC device
(DevID = 0x1480) is found, all busses in the system are considered.
Those with only SoC virtual devices are excluded, leaving only those
busses with something physically attached. This should make it
substantially easier for anyone automating use of the NBIO P-State
functionality to simply grab the PCI bus of the device in question,
without having to try to figure out which IOHC bus base to which it is
attached.

Signed-off-by: Lewis Carroll <lewis.carroll@amd.com>