Add raw HSMP interface
Add a raw HSMP binary interface to the amd_hsmp module. This allows
users to use the HSMP interface for messages on their platform that
may not be officially supported by the driver.

NOTE: This interface does not (and cannot) provide operation atomicity.
Users beware, the results returned are from the last HSMP request
written to the interface. Multiple userspace users may get incorrect
results.

Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>