Correct IOHC0 device pointer init for sockets
The current setting of the socket pci device pointer for IOHC0
makes an assumption that the order in which the northbridge
devices are sorted is IOHC 0 - 3 for each socket. Where this is
done they are not sorted at all.

Correct this by just by retrieving the IOHC0 device for
each socket using node_to_amd_nb() which returns the northbridge
device having nb->root being IOHC0.

Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>