hsmp_driver: Remove 4 IOHC devices per socket restriction
Some platforms do not adhere to the assumption that there are four
IOHC devices per socket as the code expects. Update the code so that
the assumption is removed and we calculate the number of IOHC devices
per socket based on probing of devices.

Update also include updates to the test script and kernel module
to correctly test the per-bus sysfs files.

Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>