Only track NBIO devices
Remove support for tracking PCIe busses with external devices
attached and creating sysfs entries for them. When setting
the NBIO P-state for a bus, the HSMP call is done on the NBIO
device that contains the bus, not the PCIe bus directly.

Taking this out also removes the extra overhead of tracking
and sorting pci busses, and the ongoing maintenance to keep
the soc device list up to date.

Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>