Merge smu_pci_[read|write] routines
The routines to read and write smu values are nearly identical, merge
these to routines to a single smu_rdwr() routine.

This update also allows us to add front-ends to the new smu_rdwr()
for smu_read() that validates the socket id and takes the socket
lock instead of having every caller do so.

Additional front-ends are added for hsmp_read() and hsmp_write()
to add common variables (hsmp_port) to the smu_rdwr() call.

There should be no functional changes.

Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>