Sysfs updates
In order to reduce redundant code in sysfs file show routines
and to make sysfs file creation/deletion nicer several updates
are made to the sysfs hsndling.

A hsmp_socket_attr_show define is introduced to make defining a
per-socket sysfs attr show routine that displays a single value
easier to define.

A ddr_attr_show define is added to reduce the common code in
the ddr attr show routines.

The sysfs files created by the amd_hsmp driver are coalesced
into attribute groups so we can create/delete groups of files
with a single call instead of individually.

Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>