Allow driver to load on unsupported versions
There is no reason the amd_hsmp driver cannot load and function
on systems that have a higher HSMP protocol version than what
the driver supports. The driver can still provide access to HSMP
functions that are supported by the driver.

Update the protocol version checking to only fail loading for
an invalid protocol version. When trying to load on a system
with a higher HSMP protocol version than what the driver supports
output a warning indicating the current protocol version
and what the driver supports.

Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>