hsmp_probe() fixup
There is no need to continue probing HSMP after we receive the
first error during probe, we're going to return failure anyway.
Update the probe routine to return failure when a failure is
encountered.

Additionally, the printing of the HSMP protocol version and
SMN version, plus the protocol warning, are moved until the
driver successfully is initialized. No need to print these if
the driver fails to load.

Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>