amd_hsmp.c: Emit friendly error on fam 17h
If family 17h support is not enabled and the driver is loaded on a
family 17h part, the driver does not emit any error message into the
kernel log - it only emits the driver version string. Add an error
stating that this CPU does not support HSMP.

Also remove redundant test for unsupported protocol version < 1. If
the protocol version were to return 0, the HSMP part of the SMU
firmware would be sufficiently broken that the test message would
have failed before reaching this point.

Signed-off-by: Lewis Carroll <lewis.carroll@amd.com>